+++ libs/config.json
@@ -1,0 +1,12 @@
+{
+    "loglevel": "INFO",
+    "engine_depth": 10,
+    "engine_time": 10,
+    "engine_threads": 2,
+    "engine_hash": 64,
+    "show_wdl": false,
+    "enable_chessdb": false,
+    "chessdb_timeout": 5,
+    "timer_interval": 100,
+    "confirm_interval": 100
+}
