--- server/.gitignore
@@ -1,3 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
