--- server/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "chessboard"
-version = "0.1.0"
-description = "a chinese chess learn tools"
-authors = ["atopx"]
-edition = "2021"
-
-
-[build-dependencies]
-tauri-build = { version = "1", features = [] }
-
-
-[dependencies]
-tauri = { version = "1", features = [ "shell-open"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-ort = { git = "https://github.com/pykeio/ort.git", tag = "v2.0.0-alpha.4", features = ["coreml", "load-dynamic"] }
-xcap = "0.0.10"
-ndarray = "0.15"
-tokio = { version = "1.36.0", features = ["full"] }
-reqwest = { version = "0.11", features = ["blocking"] }
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-tokio-cron-scheduler = "0.10.0"
-lazy_static = "1.4.0"
-
-[features]
-# This feature is used for production builds or when a dev server is not specified, DO NOT REMOVE!!
-custom-protocol = ["tauri/custom-protocol"]
