--- server/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    println!("cargo:rustc-link-arg=-fapple-link-rtlib");
-    tauri_build::build()
-}
