--- server/rustfmt.toml
@@ -1,3 +1,0 @@
-newline_style = "Unix"
-max_width = 106
-use_small_heuristics = "Max"
