--- server/tauri.conf.json
@@ -1,51 +1,0 @@
-{
-  "build": {
-    "beforeDevCommand": "pnpm dev",
-    "beforeBuildCommand": "pnpm build",
-    "devPath": "http://localhost:1420",
-    "distDir": "../dist"
-  },
-  "package": {
-    "productName": "chessboard",
-    "version": "0.0.0"
-  },
-  "tauri": {
-    "allowlist": {
-      "all": false,
-      "shell": {
-        "all": false,
-        "open": true
-      }
-    },
-    "windows": [
-      {
-        "title": "中国象棋",
-        "width": 660,
-        "height": 580,
-        "hiddenTitle": true,
-        "resizable": false,
-        "fullscreen": false,
-        "maximizable": false
-      }
-    ],
-    "security": {
-      "csp": null
-    },
-    "bundle": {
-      "resources": [
-        "../libs/*",
-        "../libs/config.json"
-      ],
-      "active": true,
-      "targets": "all",
-      "identifier": "top.itmeng.chessboard",
-      "icon": [
-        "icons/32x32.png",
-        "icons/128x128.png",
-        "icons/128x128@2x.png",
-        "icons/icon.icns",
-        "icons/icon.ico"
-      ]
-    }
-  }
-}
