+++ src-tauri/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "xqlink"
+version = "0.1.0"
+description = "A Tauri App"
+authors = ["you"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+# The `_lib` suffix may seem redundant but it is necessary
+# to make the lib name unique and wouldn't conflict with the bin name.
+# This seems to be only an issue on Windows, see https://github.com/rust-lang/cargo/issues/8519
+name = "xqlink_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[build-dependencies]
+tauri-build = { version = "2", features = [] }
+
+[dependencies]
+tauri = { version = "2", features = [] }
+tauri-plugin-opener = "2"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+ort = { git = "https://github.com/pykeio/ort.git", tag = "v2.0.0-rc.9" }
+xcap = { version = "0.5"}
+tracing = "0.1"
+tracing-subscriber = "0.3"
+tokio-cron-scheduler = "0.13"
+ndarray = "0.16"
+reqwest = { version = "0.12", features = ["json"] }
+tokio = { version = "1.44", default-features = false, features = [
+    "macros",
+    "rt-multi-thread",
+] }
+lazy_static = "1.5.0"
+
+[features]
+cuda = ["ort/cuda"]
+coreml = ["ort/coreml"]
+load-dynamic = ["ort/load-dynamic"]
+
+[profile.release]
+codegen-units = 1
+lto = "fat"
+panic = "abort"
