+++ src-tauri/src/main.rs
@@ -1,0 +1,4 @@
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
+
+fn main() { xqlink_lib::run() }
