+++ src/main.ts
@@ -4,3 +4,4 @@ import "./styles.css";
 
 
 createApp(App).mount("#app");
+
