+++ vite.config.ts
@@ -1,5 +1,8 @@
-import vue from "@vitejs/plugin-vue";
 import { defineConfig } from "vite";
+import vue from "@vitejs/plugin-vue";
+
+// @ts-expect-error process is a nodejs global
+const host = process.env.TAURI_DEV_HOST;
 
 // https://vitejs.dev/config/
 export default defineConfig(async () => ({
@@ -13,9 +16,17 @@ export default defineConfig(async () => ({
   server: {
     port: 1420,
     strictPort: true,
+    host: host || false,
+    hmr: host
+      ? {
+          protocol: "ws",
+          host,
+          port: 1421,
+        }
+      : undefined,
     watch: {
       // 3. tell vite to ignore watching `src-tauri`
-      ignored: ["**/server/**"],
+      ignored: ["**/src-tauri/**"],
     },
   },
 }));
